python/example_code/bedrock-runtime/cross-model-scenarios/long-term-memory-demo/requirements.txt
boto3==1.34.137
botocore==1.34.137
jmespath==1.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
